+++ src/index.ts
@@ -1,10 +1,9 @@
-import Router from './router/router';
 import WebComponent from './component/web_component';
 
 export * as i18n from './intl/i18n';
-export { translate } from './intl/i18n';
-
+export * as router from './router/router';
 export { html } from './template/template';
+export { translate } from './intl/i18n';
 export { createState } from './state/state';
+export { WebComponent };
 export { defineWebComponent } from './component/web_component';
-export { Router, WebComponent };